Augment "vue" module instead of "@vue/runtime-core"
When using a package manager stricter than npm such as pnpm [0], @vue/runtime-core
is not directly accessible and therefore cannot be augmented.

This contribution updates the module augmentation to use "vue" (which is
what this package define as a peer dependency). This is also what the
current Vue documentation suggests to use [1].

[0] https://pnpm.io/motivation#creating-a-non-flat-node_modules-directory
[1] https://vuejs.org/api/utility-types.html#componentcustomproperties